+++ src/DefaultServerProblemDetails/DefaultServerProblemDetailSuperType.ts
@@ -1,0 +1,8 @@
+import { ProblemDetails } from "@tectonique/api-standards";
+import DefaultServerProblemDetailsCollection from "./DefaultServerProblemDetailsCollection";
+
+type DefaultServerProblemDetailSuperType = ProblemDetails.infer<
+  typeof DefaultServerProblemDetailsCollection
+>;
+
+export default DefaultServerProblemDetailsCollection;