+++ src/DefaultServerProblemDetails/DefaultServerProblemDetailsCollection.ts
@@ -1,0 +1,6 @@
+import * as problemDetails from "./problemDetails";
+import { ProblemDetails } from "@tectonique/api-standards";
+
+const DefaultServerProblemDetailsCollection = { ...problemDetails };
+
+export default DefaultServerProblemDetailsCollection;