+++ src/DefaultServerProblemDetails/index.ts
@@ -1,0 +1,3 @@
+export * from "./problemDetails";
+export { default as DefaultServerProblemDetailsCollection } from "./DefaultServerProblemDetailsCollection";
+export { default as DefaultServerProblemDetailSuperType } from "./DefaultServerProblemDetailSuperType";