+++ src/Envelopes/index.ts
@@ -1,5 +1,4 @@
 import { INestApplication } from "@nestjs/common";
-import { FallbackInternalServerProblemDetailGenerator } from "../types";
 import ProblemDetailHandlingExceptionFilter from "./ProblemDetailHandlingExceptionFilter";
 import { HttpAdapterHost } from "@nestjs/core";
 import ResponseEnvelopeWrappingInterceptor from "./ResponseEnvelopeWrappingInterceptor";
@@ -7,18 +6,9 @@ import ResponseEnvelopeWrappingInterceptor from "./ResponseEnvelopeWrappingInter
 export { default as ProblemDetailHandlingExceptionFilter } from "./ProblemDetailHandlingExceptionFilter";
 export { default as ResponseEnvelopeWrappingInterceptor } from "./ResponseEnvelopeWrappingInterceptor";
 
-export function useEnvelopeAndProblemDetailHandlers(
-  app: INestApplication,
-  fallbackInternalServerProblemDetailGenerator: FallbackInternalServerProblemDetailGenerator
-) {
+export function useEnvelopeAndProblemDetailHandlers(app: INestApplication) {
   const { httpAdapter } = app.get(HttpAdapterHost);
 
-  app.useGlobalFilters(
-    new ProblemDetailHandlingExceptionFilter(
-      httpAdapter,
-      fallbackInternalServerProblemDetailGenerator
-    )
-  );
-
+  app.useGlobalFilters(new ProblemDetailHandlingExceptionFilter(httpAdapter));
   app.useGlobalInterceptors(new ResponseEnvelopeWrappingInterceptor());
 }