+++ src/types.ts
@@ -1,4 +1,0 @@
-import { ProblemDetails } from "@tectonique/api-standards";
-
-export type FallbackInternalServerProblemDetailGenerator =
-  () => ProblemDetails.ProblemDetail<any, any, any>;